+++ demo/demo01/index.html
@@ -82,7 +82,7 @@
     </div>
   </div>
 </div>
-<script src="./js/fullpage.js"></script>
+<script src="../../build/fullpage.js"></script>
 <script src="./js/index.js"></script>
 </body>
 </html>
