--- demo/demo01/js/fullpage-es6.js
@@ -1,634 +1,0 @@
-/**
- * fullPage v1.3.1
- * https://github.com/kisnows/fullpage.js
- *
- * Apache License
- *
- * A light JavaScript framework for developer to build some fullPage site by a simple way.Write with pure JavaScript.
- *
- * Author: yq12315@gmail.com
- */
-
-//helper===========================================================
-
-function $$(el, parent) {
-  if (!parent) {
-    return document.querySelectorAll(el);
-  } else {
-    return parent.querySelectorAll(el);
-  }
-}
-
-function setCss(el, props) {
-  let prop;
-  for (prop in props) {
-    if (props.hasOwnProperty(prop)) {
-      el.style[prop] = props[prop];
-    }
-  }
-  return el;
-}
-
-function setAttr() {
-
-  function translate(el, value, direction) {
-    if (direction === 'y') {
-      setCss(el, {
-        'transform': "translate3d(0," + value + "px,0)",
-        '-webkit-transform': "translate3d(0," + value + "px,0)"
-      });
-      //console.log('setAttr Done');
-    } else if (direction === 'x') {
-      setCss(el, {
-        "transform": "translate3d(" + value + "px,0,0)",
-        "-webkit-transform": "translate3d(" + value + "px,0,0)"
-      });
-    }
-  }
-
-  return {
-    translate: translate
-  };
-}
-
-/**
- * 动画检测
- */
-/* From Modernizr */
-function whichTransitionEvent() {
-  let t;
-  let el = document.createElement('fakeelement');
-  let transitions = {
-    'transition': 'transitionend',
-    'OTransition': 'oTransitionEnd',
-    'MozTransition': 'transitionend',
-    'WebkitTransition': 'webkitTransitionEnd',
-    'MsTransition': 'msTransitionEnd'
-  };
-
-  for (t in transitions) {
-    if (el.style[t] !== undefined) {
-      return transitions[t];
-    }
-  }
-}
-
-/* Listen for a transition! */
-const transitionEvent = whichTransitionEvent();
-
-//end helper=========================================================
-
-const stepHeight = $$('body')[0].scrollHeight;
-const stepWidth = $$('body')[0].scrollWidth;
-
-let sectionContent;
-let sections;
-let translate3dY = 0;
-
-let options = {};
-const defaults = {
-  threshold: 50,              //触发滚动事件的阈值，越小越灵敏
-  pageSpeed: 500,             //滚屏速度，单位为毫秒 ms
-  autoScroll: false,          //DONE 是否自动播放
-  autoScrollDuration: 1000,   //DONE 自动播放间隔时间
-  loopSection: true,          //DONE Section循环滚动
-  loopSlide: true,            //DONE Slide循环滑动
-  afterLoad: null,            //DONE 页面载入事件
-  beforeLeave: null,           //DONE 页面离开事件
-  afterSlideLoad: null,        //DONE slide 载入事件
-  beforeSlideLeave: null      //DONE slide 离开事件
-};
-
-
-/**
- * 绑定触摸事件
- * @param ele {string}
- */
-function bindTouchMove(ele) {
-
-  let startPos = {},
-    movePos = {},
-    endPos = {};
-  let diffX,
-    diffY;
-  let touch;
-  let onceTouch = false;                  //判断是否为一次触摸，保证一次触摸只触发一次事件
-
-  let threshold = options.threshold;      //阈值,灵敏度，越小越灵敏
-  let isVertical;                         //是否为垂直滚动事件
-
-  function touchstartHandle(event) {
-    //onceTouch首先置为true，表明开始了一次触摸
-    onceTouch = true;
-    // 初始化 x,y 值，防止点击一次后出现假 move 事件
-    startPos = {};
-    endPos = {};
-    if (event.target.tagName.toLowerCase() !== 'a') {
-      event.preventDefault();
-    }
-    touch = event.touches[0];
-    startPos.x = touch.pageX;
-    startPos.y = touch.pageY;
-  }
-
-  function touchmoveHandle(event) {
-
-    event.preventDefault();
-    touch = event.touches[0];
-    movePos.x = touch.pageX;
-    movePos.y = touch.pageY;
-    diffX = startPos.x - movePos.x;
-    diffY = startPos.y - movePos.y;
-
-    //如果页面正在滚动或者不是一次滚动事件，则直接return掉
-    if (page.isScrolling || !onceTouch) {
-      return false;
-    }
-
-    isVertical = Math.abs(diffX) - Math.abs(diffY) <= 0;
-    //如果diff大于阈值，则事件触发，将onceTouch置为false
-    onceTouch = Math.max(diffX, diffY) <= threshold;
-    if (!isVertical) {
-      if (diffX > threshold) {
-        //Move to left
-        page.slide.next();
-      } else if (diffX < -threshold) {
-        //Move to right
-        page.slide.pre();
-      }
-    } else {
-      //isVertical = true;
-      if (diffY > threshold) {
-        //Move to top
-        page.move.next();
-      } else if (diffY < -threshold) {
-        //Move to bottom
-        page.move.pre();
-      }
-    }
-  }
-
-  function touchendHandle(event) {
-    if (event.target.tagName.toLowerCase() !== 'a') {
-      event.preventDefault();
-    }
-    //重置onceTouch为true
-    onceTouch = true;
-  }
-
-  ele.addEventListener('touchstart', touchstartHandle, false);
-
-  ele.addEventListener('touchmove', touchmoveHandle, false);
-
-  ele.addEventListener('touchend', touchendHandle, false);
-}
-
-/**
- * 绑定鼠标滚动事件
- */
-function bindMouseWheel() {
-  //FIXME change the way binding event.
-  let type;
-  let deltaY;
-
-  if (navigator.userAgent.toLowerCase().indexOf('firefox') !== -1) {
-    type = 'DOMMouseScroll';
-  } else {
-    type = 'mousewheel';
-  }
-
-  function mouseWheelHandle(event) {
-    if (page.isScrolling) {
-      return false;
-    }
-    deltaY = event.detail || -event.wheelDelta || event.deltaY;
-    if (deltaY > 0) {
-      page.move.next();
-      //console.log('next');
-    } else if (deltaY < 0) {
-      page.move.pre();
-      //console.log('pre');
-    }
-  }
-
-  document.addEventListener(type, mouseWheelHandle, false);
-}
-
-/**
- * 绑定键盘事件
- */
-function bindKeyboard() {
-  function keyboardHandle(event) {
-    let key = event.keyCode || event.which;
-    switch (key) {
-      case 37:
-        page.slide.pre();
-        break;
-      case 38:
-        page.move.pre();
-        break;
-      case 39:
-        page.slide.next();
-        break;
-      case 40:
-        page.move.next();
-        break;
-    }
-  }
-
-  document.addEventListener('keydown', keyboardHandle, false);
-}
-
-/**
- * 页面滚动主要逻辑
- * @type {{nowPage: number, scrollPage: Function, scrollSlide: Function, move: {next: Function, pre: Function}, moveTo: Function, slide: {next: Function, pre: Function}}}
- */
-let page = {
-  nowPage: 1,
-  isScrolling: false,
-  modifyIsScrolling: function () {
-
-  },
-  /**
-   * Scroll to a specified page.
-   * @param pageIndex {number} The page index you want scroll to.
-   * @returns {boolean}
-   */
-  scrollPage: function (pageIndex) {
-
-    let pageDiff = pageIndex - page.nowPage;
-    let leaveSection = sections[page.nowPage - 1];
-    let nowSection = sections[pageIndex - 1];
-    let controllers = $$('.fp-controller-dotted');
-    if (pageIndex >= 1 && pageIndex <= sections.length && !page.isScrolling && pageDiff) {
-
-      if (typeof options.beforeLeave === 'function') {
-        /**
-         * leaveSection 函数内部 this 指向，将要离开的 section
-         * page.nowPage 将要离开页面的 index
-         * pageIndex    将要载入页面的 index
-         */
-        options.beforeLeave.call(leaveSection, page.nowPage, pageIndex);
-      }
-
-      leaveSection.classList.remove('active');
-      addClassToOneEle(controllers, pageIndex - 1);
-      translate3dY -= pageDiff * stepHeight;
-      setAttr().translate(sectionContent, translate3dY, 'y');
-      page.isScrolling = true;
-      page.nowPage = pageIndex;
-      nowSection.classList.add('active');
-
-      if (typeof options.afterLoad === 'function') {
-        options.pageSpeed = options.pageSpeed ? 500 : options.pageSpeed;
-        setTimeout(function () {
-          /**
-           * nowSection 函数内部 this 指向，载入后的 section
-           * pageIndex 载入后的 index
-           */
-          options.afterLoad.call(nowSection, pageIndex);
-        }, options.pageSpeed);
-      }
-      return true;
-    } else {
-      return false;
-    }
-  },
-  /**
-   * Scroll to a specified slide.
-   * @param slideIndex {number} The slide index you want scroll to.
-   * @returns {boolean}
-   */
-  scrollSlide: function (slideIndex) {
-
-    //获取slide包裹层
-    let slideWrap = $$('.fp-slide-wrap', sections[page.nowPage - 1])[0];
-
-    if (!slideWrap) {
-      console.log('This page has no slide');
-      return false;
-    }
-
-    //当前页面下所有的slide
-    let slide = sections[page.nowPage - 1].querySelectorAll('.fp-slide');
-
-    //当前页面上存储的数据
-    let slideData = slideWrap.dataset;
-
-    //当前页面上slide的index
-    let slideNowIndex = parseInt(slideData.index);
-
-    //当前页面上slide的x轴偏移值
-    let slideX = slideData.x;
-
-    let slideDiff = slideIndex - slideNowIndex;
-
-    if (slideIndex >= 1 && slideIndex <= slide.length && !page.isScrolling) {
-      if (typeof options.beforeSlideLeave === 'function') {
-        /**
-         * leaveSlide           函数内部 this 指向，将要离开的 slide
-         * page.nowPage         将要离开 section 的 index
-         * slideNowIndex        将要离开 slide 的 index
-         * slideIndex           将要载入 slide 的 index
-         */
-        options.beforeSlideLeave.call(slide[slideNowIndex - 1], page.nowPage, slideNowIndex, slideIndex);
-      }
-
-      slide[slideNowIndex - 1].classList.remove('active');
-      slideX -= slideDiff * stepWidth;
-      setAttr().translate(slideWrap, slideX, 'x');
-      page.isScrolling = true;
-      slideData.x = slideX;
-      slideData.index = slideIndex;
-      slide[slideIndex - 1].classList.add('active');
-
-      if (typeof options.afterSlideLoad === 'function') {
-        options.pageSpeed = options.pageSpeed ? 500 : options.pageSpeed;
-        setTimeout(function () {
-          /**
-           * nowSection 函数内部 this 指向，载入后的 section
-           * pageIndex 载入后的 index
-           */
-          options.afterSlideLoad.call(slide[slideIndex - 1], page.nowPage, slideIndex);
-        }, options.pageSpeed);
-      }
-      return true;
-    }
-    return false;
-  },
-  /**
-   * Scroll to a specified section and slide.
-   * @param pageIndex {number}
-   * @param slideIndex {number}
-   * @returns {boolean}
-   */
-  moveTo: function (pageIndex, slideIndex) {
-    //DONE move to a specify section or slide
-    if (page.scrollPage(pageIndex)) {
-      if (slideIndex) {
-        //DONE move to a specify slide
-        return !!page.scrollSlide(slideIndex);
-      }
-      return true;
-    } else {
-      return false;
-    }
-  },
-  move: {
-    next: function (callback) {
-
-      if (page.scrollPage(page.nowPage + 1)) {
-
-        let arg = Array.prototype.slice.call(arguments, 1);
-
-        if (typeof callback === 'function') {
-          callback.call(null, arg);
-        }
-        return true;
-      } else if (options.loopSection) {
-
-        page.moveTo(1);
-
-        return true;
-      } else {
-        return false;
-      }
-    },
-    pre: function (callback) {
-
-      if (page.scrollPage(page.nowPage - 1)) {
-
-        let arg = Array.prototype.slice.call(arguments, 1);
-
-        if (typeof callback === 'function') {
-          callback.call(null, arg);
-        }
-        return true;
-      } else {
-        return false;
-      }
-    }
-  },
-  slide: {
-    /**
-     * slide move 方法，移动到上一个或下一个 slide
-     * @param {string} direction 要移动的方向，next 为下一个， pre 为上一个
-     * @returns {boolean}
-     */
-    move: function (direction) {
-      let slideWrap = $$('.fp-slide-wrap', sections[page.nowPage - 1])[0];
-      let slide = sections[page.nowPage - 1].querySelectorAll('.fp-slide');
-      // slideNowIndexChange slideNowIndex 将要的变化
-      let slideNowIndexChange;
-      // slideWillBe 将要滚到slide的index
-      let slideWillBe;
-      if (direction === 'next') {
-        slideNowIndexChange = 1;
-        slideWillBe = 1;
-      } else {
-        slideNowIndexChange = -1;
-        slideWillBe = slide.length;
-      }
-      if (!slideWrap) {
-        return false;
-      } else {
-        let slideData = slideWrap.dataset;
-        let slideNowIndex = parseInt(slideData.index);
-
-        if (page.scrollSlide(slideNowIndex + slideNowIndexChange)) {
-
-          slideData.index = slideNowIndex + slideNowIndexChange;
-          return true;
-        } else if (options.loopSlide && page.scrollSlide(slideWillBe)) {
-          slideData.index = slideWillBe;
-          return true;
-        }
-        return false;
-      }
-    },
-    next: function () {
-      this.move('next');
-    },
-    pre: function () {
-      this.move('pre');
-    }
-  }
-};
-
-function pageController() {
-  function init() {
-    createControllerNode();
-    bindEvent();
-  }
-
-  function createControllerNode() {
-    let controllerWrap = document.createElement('div');
-    let controllerText = '';
-    controllerWrap.className = 'fp-controller';
-    for (let i = sections.length; i--; i > 0) {
-      controllerText += '<div class="fp-controller-dotted"></div>';
-    }
-    controllerWrap.innerHTML = controllerText;
-    document.body.appendChild(controllerWrap);
-  }
-
-  function bindEvent() {
-    let controllers = $$('.fp-controller-dotted');
-    for (let i = controllers.length - 1; i >= 0; i--) {
-      controllers[i].addEventListener('click', helper(i + 1), false)
-    }
-    function helper(i) {
-      return function () {
-        addClassToOneEle(controllers, i - 1);
-        page.moveTo(i);
-      }
-    }
-  }
-
-  return init();
-}
-
-/**
- * 只给一组元素中的某一个元素添加class
- * @param els 一组元素
- * @param theOne 要添加元素的index值
- */
-function addClassToOneEle(els, theOne) {
-  for (let j = els.length - 1; j >= 0; j--) {
-    els[j].classList.remove('active');
-  }
-  els[theOne].classList.add('active');
-}
-
-/**
- * 初始化页面主体元素
- * @returns {{init: init, initContent: initContent, initSlide: initSlide}}
- */
-function initEle() {
-
-  function init() {
-    initContent();
-    initSlide();
-    pageController();
-    initController();
-  }
-
-  /**
-   * 初始化 Section
-   */
-  function initContent() {
-    setCss(sectionContent, {
-      "transform": "translate3d(0,0,0)",
-      "-webkit-transform": "translate3d(0,0,0)",
-      "transitionDuration": options.pageSpeed + 'ms',
-      "-webkit-transitionDuration": options.pageSpeed + 'ms',
-      "display": "block"
-    });
-    sectionContent.addEventListener(transitionEvent, function () {
-      page.isScrolling = false;
-    }, false);
-    for (let i = sections.length - 1; i >= 0; i--) {
-      sections[i].style.height = stepHeight + 'px';
-    }
-    sections[page.nowPage - 1].classList.add('active');
-  }
-
-  /**
-   * 初始化 Slide
-   */
-  function initSlide() {
-    let slideWrap = $$('.fp-slide-wrap');
-    let slides;
-
-    function slideWrapInitHandle() {
-      page.isScrolling = false;
-    }
-
-    for (let i = slideWrap.length - 1; i >= 0; i--) {
-      slides = $$('.fp-slide', slideWrap[i]);
-      for (let j = slides.length - 1; j >= 0; j--) {
-        slides[j].style.width = stepWidth + 'px';
-      }
-      slideWrap[i].style.width = slides.length * stepWidth + 'px';
-      slideWrap[i].dataset.x = '0';
-      slideWrap[i].dataset.index = '1';
-      slideWrap[i].addEventListener(transitionEvent, slideWrapInitHandle, false);
-    }
-  }
-
-  function initController() {
-    let controllers = $$('.fp-controller-dotted');
-    controllers[page.nowPage - 1].classList.add('active');
-  }
-
-  return {
-    init: init,
-    initContent: initContent,
-    initSlide: initSlide
-  };
-}
-
-/**
- * 初始化定制内容
- * @returns {{init: init}}
- */
-function initProp() {
-
-  function init() {
-    for (let key in prop) {
-      if (prop.hasOwnProperty(key)) {
-        prop[key]();
-      }
-    }
-  }
-
-  let prop = {
-    autoScroll: function () {
-      let timer = null;
-      if (options.autoScroll) {
-        timer = setInterval(function () {
-          page.move.next();
-        }, options.autoScrollDuration);
-      }
-    }
-  };
-
-  return {
-    init: init
-  };
-}
-
-/**
- * 注册事件
- * @param ele {String} 要绑定的元素
- */
-function bindEvent(ele) {
-  bindTouchMove(ele);
-  bindKeyboard();
-  bindMouseWheel();
-}
-
-function init(ele, Customize) {
-
-  sectionContent = $$(ele)[0];
-  sections = $$('.fp-section');
-  options = Object.assign({}, defaults, Customize);
-
-  initEle().init();
-  initProp().init();
-  bindEvent(sectionContent);
-}
-
-//API
-let fullpage = {
-  init: init,
-  scrollPage: page.scrollPage,
-  scrollSlide: page.scrollSlide,
-  moveTo: page.moveTo,
-  moveToNext: page.move.next,
-  moveToPre: page.move.pre,
-  slideToNext: page.slide.next,
-  slideToPre: page.slide.pre
-};
-
-export default fullpage;
